requirements.txt
Django==5.1.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-debug-toolbar==5.0.1
django-modeltranslation==0.19.9
django-import-export==4.2.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
python-dotenv==1.0.1
psycopg2==2.9.10
